Rename "build" module to "builder"
Given we have various functions called "build", and in any case, that's a verb so it would make sense to associate it with an active action. "Builder" is a noun, which better represents an object that can be imported.
